Results page display enhancements per Andrew@RFID
The system now displays the following:
Finish (TOD): The time of day of the finish
Start (TOD): Assigned start time
Wave (TOD): Wave start time for the participants assigned wave
Start (Last Seen): The last time we see the runner at the start line. It should be equal to the Start (TOD) column.
Start Offset: The difference between the Start (Last Seen) and the Wave (TOD) columns. 

If the last seen is blank (e.g. no chip reads at all at the start, the offset becomes zero.  If there is an override for the start time, then the offset is set to zero and the last seen is set to the override value.

All times are also now shown in HH:MM:SS.sss format